refactor: move loader to "options" folder
